feat: refactor keras agents
